+++ src/main/java/ru/sbt/mipt/oop/door/Door.java
@@ -1,4 +1,4 @@
-package ru.sbt.mipt.oop.room;
+package ru.sbt.mipt.oop.door;
 
 public class Door {
     private final String id;
