+++ src/main/java/ru/sbt/mipt/oop/notifications/Notifier.java
@@ -1,6 +1,6 @@
 package ru.sbt.mipt.oop.notifications;
 
-import ru.sbt.mipt.oop.room.Door;
+import ru.sbt.mipt.oop.door.Door;
 import ru.sbt.mipt.oop.light.Light;
 import ru.sbt.mipt.oop.Room;
